More reliable mouse position reporting
Depending on the structure of the page, the computed mouse position was
not correct. Typically the case happend when there is no offset between
the canvas and the view area, but an offset on the view area.

The MouseEvent.offsetX and offsetY functions are now widely enough spread
to use them and avoid complex computations.